uv/tests: move ecosystem project copying to TestContext
So that we can easily reuse ecosystem projects in other tests.
